Feature/login

+++ src/main/java/com/nexters/wiw/api/service/UserService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import com.nexters.wiw.api.domain.User;
 import com.nexters.wiw.api.domain.UserRepository;
 import com.nexters.wiw.api.domain.error.ErrorType;
+import com.nexters.wiw.api.exception.UnAuthorizedException;
 import com.nexters.wiw.api.exception.UserNotExistedException;
 import com.nexters.wiw.api.ui.UserRequestDto;
 
@@ -13,22 +14,30 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import lombok.extern.slf4j.Slf4j;
-
 @Service
 public class UserService {
+
+    @Autowired
+    private AuthService authService;
+
     @Autowired
     private UserRepository userRepository;
 
     @Autowired
     private PasswordEncoder bCryptPasswordEncoder;
 
-    public User getOne(Long id) {
+    public User getOne(String authHeader, Long id) {
+        if (!authService.isValidateToken(authHeader))
+            throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "UNAUTHORIZED");
+
         return userRepository.findById(id)
                 .orElseThrow(() -> new UserNotExistedException(ErrorType.NOT_FOUND, "NOT_FOUND"));
     }
 
-    public List<User> getList(String email, String nickname) {
+    public List<User> getList(String authHeader, String email, String nickname) {
+        if (!authService.isValidateToken(authHeader))
+            throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "UNAUTHORIZED");
+
         return userRepository.findByEmailOrNickname(email, nickname);
     }
 
@@ -39,12 +48,18 @@ public class UserService {
     }
 
     @Transactional
-    public User patch(final Long id, final User user) {
-        return getOne(id).update(user);
+    public User patch(String authHeader, final Long id, final User user) {
+        if (!authService.isValidateToken(authHeader))
+            throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "UNAUTHORIZED");
+
+        return getOne(authHeader, id).update(user);
     }
 
     @Transactional
-    public void delete(final Long id) {
+    public void delete(String authHeader, final Long id) {
+        if (!authService.isValidateToken(authHeader))
+            throw new UnAuthorizedException(ErrorType.UNAUTHORIZED, "UNAUTHORIZED");
+
         userRepository.deleteById(id);
     }
 }
